+++ .projenrc.ts
@@ -1,10 +1,10 @@
 import { typescript } from "projen";
-import { PackageVersions, CI_Versions } from "./const";
 import { ConfigureTasks } from "./projenrc/ConfigureTasks";
 import { CdktfWorkflows } from "./projenrc/workflows/CdktfWorkflows";
 import { K8sValidate } from "./projenrc/workflows/K8sValidate";
 import { UpgradeCharts } from "./projenrc/workflows/UpgradeCharts";
 import { Upgradepackages } from "./projenrc/workflows/UpgradePackages";
+import { PackageVersions, CI_Versions } from "./src/const";
 
 const project = new typescript.TypeScriptAppProject({
   defaultReleaseBranch: "master",
@@ -57,7 +57,7 @@ const project = new typescript.TypeScriptAppProject({
   eslint: true,
   eslintOptions: {
     prettier: true,
-    dirs: ["src", "projen"],
+    dirs: ["src"],
   },
 
   deps: [
