+++ projenrc/workflows/CdktfWorkflows.ts
@@ -1,7 +1,7 @@
 import { typescript } from "projen";
 import { GithubWorkflow } from "projen/lib/github";
 import { JobPermission } from "projen/lib/github/workflows-model";
-import { CI_Versions } from "../../const";
+import { CI_Versions } from "../../src/const";
 
 export function CdktfWorkflows(project: typescript.TypeScriptAppProject) {
   for (const context of ["build", "deploy"]) {
